requirements.txt
arrow==1.3.0
asgiref==3.7.2
autopep8==2.0.4
blessed==1.20.0
cached-property==1.5.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
coverage==7.3.2
crispy-bootstrap5==0.7
distlib==0.3.7
Django==4.2.7
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.4
django-picklefield==3.1
django-q==1.3.9
django-silk==5.0.4
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.12.0
filelock==3.12.4
gprof2dot==2022.7.29
identify==2.5.30
idna==3.4
Markdown==3.5.1
mock==5.1.0
nodeenv==1.8.0
parameterized==0.9.0
Pillow==10.1.0
platformdirs==3.11.0
pre-commit==3.5.0
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.6
wcwidth==0.2.8
